Bring in stream metadata from static CSVs

+++ core/src/test/java/uk/ac/ucl/rits/inform/datasinks/emapstar/waveform/TestWaveformProcessing.java
@@ -48,7 +48,7 @@ class TestWaveformProcessing extends MessageProcessingBase {
         String sourceStreamId;
         String mappedStreamName;
         int numSamples;
-        long samplingRate;
+        int samplingRate;
         int maxSamplesPerMessage;
         String sourceLocation;
         String mappedLocation;
